[when] move MLIR compile logic to separate file
